Code optimizations and fixes

+++ src/test_signals.cpp
@@ -55,7 +55,7 @@ void on_process(void* userdata, spa_io_position* position) {
       case TestSignalType::sine_wave: {
         d->ts->sine_phase += 2.0F * std::numbers::pi_v<float> * d->ts->sine_frequency / static_cast<float>(rate);
 
-        signal = 0.5F * sinf(d->ts->sine_phase);
+        signal = 0.5F * std::sin(d->ts->sine_phase);
 
         break;
       }
